Added template support
