+++ .vscode/settings.json
@@ -2,5 +2,6 @@
 	"editor.tabSize": 2,
 	"typescript.tsdk": "node_modules/typescript/lib",
 	"typescript.enablePromptUseWorkspaceTsdk": true,
-	"css.lint.unknownAtRules": "ignore"
+	"css.lint.unknownAtRules": "ignore",
+	"prettier.enable": false
 }
