+++ src/components/LogoFull.tsx
@@ -2,10 +2,11 @@ import React from 'react';
 
 export const LogoFull: React.FC<{
 	color: string;
+	secondaryColor?: string;
 	wave?: number;
 	width?: number;
 	height?: number;
-}> = ({color, wave = 0, width = 600, height = 150}) => {
+}> = ({color, secondaryColor = "#000000", wave = 0, width = 600, height = 150}) => {
 	return (
 		<div>
 			<svg
@@ -23,29 +24,29 @@ export const LogoFull: React.FC<{
 					<clipPath id="8ade551c01">
 						<path
 							d="M 25.605469 15 L 146 15 L 146 135 L 25.605469 135 Z M 25.605469 15 "
-							clip-rule="nonzero"
+							clipRule="nonzero"
 						/>
 					</clipPath>
 				</defs>
-				<g clip-path="url(#8ade551c01)">
+				<g clipPath="url(#8ade551c01)">
 					<path
 						fill={color}
 						d="M 145.605469 75 C 145.605469 76.964844 145.507812 78.925781 145.316406 80.882812 C 145.125 82.835938 144.835938 84.777344 144.453125 86.707031 C 144.070312 88.632812 143.589844 90.535156 143.019531 92.417969 C 142.449219 94.296875 141.789062 96.144531 141.039062 97.960938 C 140.285156 99.777344 139.445312 101.550781 138.519531 103.285156 C 137.59375 105.015625 136.585938 106.699219 135.492188 108.335938 C 134.402344 109.96875 133.230469 111.542969 131.984375 113.0625 C 130.738281 114.582031 129.421875 116.035156 128.03125 117.425781 C 126.640625 118.816406 125.1875 120.132812 123.667969 121.378906 C 122.148438 122.628906 120.574219 123.796875 118.9375 124.886719 C 117.304688 125.980469 115.621094 126.988281 113.890625 127.914062 C 112.15625 128.839844 110.382812 129.679688 108.566406 130.433594 C 106.75 131.183594 104.902344 131.847656 103.023438 132.417969 C 101.140625 132.988281 99.238281 133.464844 97.308594 133.847656 C 95.382812 134.230469 93.441406 134.519531 91.484375 134.710938 C 89.53125 134.902344 87.570312 135 85.605469 135 C 83.640625 135 81.679688 134.902344 79.722656 134.710938 C 77.769531 134.519531 75.828125 134.230469 73.898438 133.847656 C 71.972656 133.464844 70.070312 132.988281 68.1875 132.417969 C 66.308594 131.847656 64.460938 131.183594 62.644531 130.433594 C 60.828125 129.679688 59.054688 128.839844 57.320312 127.914062 C 55.589844 126.988281 53.90625 125.980469 52.269531 124.886719 C 50.636719 123.796875 49.058594 122.628906 47.542969 121.378906 C 46.023438 120.132812 44.566406 118.816406 43.179688 117.425781 C 41.789062 116.035156 40.472656 114.582031 39.222656 113.0625 C 37.976562 111.542969 36.808594 109.96875 35.71875 108.335938 C 34.625 106.699219 33.617188 105.015625 32.691406 103.285156 C 31.761719 101.550781 30.925781 99.777344 30.171875 97.960938 C 29.421875 96.144531 28.757812 94.296875 28.1875 92.417969 C 27.617188 90.535156 27.140625 88.632812 26.757812 86.707031 C 26.375 84.777344 26.085938 82.835938 25.894531 80.882812 C 25.703125 78.925781 25.605469 76.964844 25.605469 75 C 25.605469 73.035156 25.703125 71.074219 25.894531 69.117188 C 26.085938 67.164062 26.375 65.222656 26.757812 63.292969 C 27.140625 61.367188 27.617188 59.464844 28.1875 57.582031 C 28.757812 55.703125 29.421875 53.855469 30.171875 52.039062 C 30.925781 50.222656 31.761719 48.449219 32.691406 46.714844 C 33.617188 44.984375 34.625 43.300781 35.71875 41.664062 C 36.808594 40.03125 37.976562 38.457031 39.222656 36.9375 C 40.472656 35.417969 41.789062 33.964844 43.179688 32.574219 C 44.566406 31.183594 46.023438 29.867188 47.542969 28.621094 C 49.058594 27.371094 50.636719 26.203125 52.269531 25.113281 C 53.90625 24.019531 55.589844 23.011719 57.320312 22.085938 C 59.054688 21.160156 60.828125 20.320312 62.644531 19.566406 C 64.460938 18.816406 66.308594 18.152344 68.1875 17.582031 C 70.070312 17.011719 71.972656 16.535156 73.898438 16.152344 C 75.828125 15.769531 77.769531 15.480469 79.722656 15.289062 C 81.679688 15.097656 83.640625 15 85.605469 15 C 87.570312 15 89.53125 15.097656 91.484375 15.289062 C 93.441406 15.480469 95.382812 15.769531 97.308594 16.152344 C 99.238281 16.535156 101.140625 17.011719 103.023438 17.582031 C 104.902344 18.152344 106.75 18.816406 108.566406 19.566406 C 110.382812 20.320312 112.15625 21.160156 113.890625 22.085938 C 115.621094 23.011719 117.304688 24.019531 118.9375 25.113281 C 120.574219 26.203125 122.148438 27.371094 123.667969 28.621094 C 125.1875 29.867188 126.640625 31.183594 128.03125 32.574219 C 129.421875 33.964844 130.738281 35.417969 131.984375 36.9375 C 133.230469 38.457031 134.402344 40.03125 135.492188 41.664062 C 136.585938 43.300781 137.59375 44.984375 138.519531 46.714844 C 139.445312 48.449219 140.285156 50.222656 141.039062 52.039062 C 141.789062 53.855469 142.449219 55.703125 143.019531 57.582031 C 143.589844 59.464844 144.070312 61.367188 144.453125 63.292969 C 144.835938 65.222656 145.125 67.164062 145.316406 69.117188 C 145.507812 71.074219 145.605469 73.035156 145.605469 75 Z M 145.605469 75 "
-						fill-opacity="1"
-						fill-rule="nonzero"
+						fillOpacity="1"
+						fillRule="nonzero"
 					/>
 				</g>
 				<path
-					fill="#000000"
+					fill={secondaryColor}
 					d="M 78.878906 38.238281 L 92.023438 38.238281 L 75.757812 76.574219 L 75.753906 76.582031 L 60.976562 111.738281 L 47.835938 111.738281 Z M 99.683594 46.066406 L 130.382812 76.78125 L 74.988281 104.617188 L 82.085938 87.703125 L 110.773438 74.117188 L 94.679688 57.792969 Z M 99.683594 46.066406 "
-					fill-opacity="1"
-					fill-rule="evenodd"
+					fillOpacity="1"
+					fillRule="evenodd"
 				/>
 				<path
 					fill={color}
 					d="M 172.035156 97.011719 L 172.035156 56.8125 L 199.554688 56.8125 L 199.554688 65.71875 L 182.671875 65.71875 L 182.671875 73.386719 L 197.144531 73.386719 L 197.144531 82.292969 L 182.671875 82.292969 L 182.671875 97.011719 Z M 206.890625 97.011719 L 206.890625 56.8125 L 221.855469 56.8125 C 224.742188 56.8125 227.402344 57.203125 229.835938 57.988281 C 232.265625 58.730469 234.226562 60.050781 235.710938 61.945312 C 237.195312 63.84375 237.9375 66.460938 237.9375 69.800781 C 237.9375 73.015625 237.195312 75.632812 235.710938 77.652344 C 234.226562 79.675781 232.265625 81.160156 229.835938 82.105469 C 227.402344 83.054688 224.742188 83.527344 221.855469 83.527344 L 217.527344 83.527344 L 217.527344 97.011719 Z M 217.527344 75.117188 L 221.113281 75.117188 C 223.21875 75.117188 224.804688 74.664062 225.875 73.757812 C 226.988281 72.851562 227.546875 71.53125 227.546875 69.800781 C 227.546875 68.066406 226.988281 66.871094 225.875 66.214844 C 224.804688 65.554688 223.21875 65.222656 221.113281 65.222656 L 217.527344 65.222656 Z M 219.9375 80.746094 L 227.113281 73.695312 L 240.164062 97.011719 L 228.289062 97.011719 Z M 258.9375 97.753906 C 255.679688 97.753906 252.816406 96.96875 250.34375 95.402344 C 247.867188 93.835938 245.929688 91.507812 244.527344 88.414062 C 243.167969 85.28125 242.488281 81.386719 242.488281 76.726562 C 242.488281 72.191406 243.167969 68.398438 244.527344 65.347656 C 245.929688 62.253906 247.867188 59.945312 250.34375 58.421875 C 252.816406 56.855469 255.679688 56.070312 258.9375 56.070312 C 262.195312 56.070312 265.0625 56.855469 267.535156 58.421875 C 270.007812 59.945312 271.925781 62.253906 273.285156 65.347656 C 274.6875 68.398438 275.390625 72.191406 275.390625 76.726562 C 275.390625 81.386719 274.6875 85.28125 273.285156 88.414062 C 271.925781 91.507812 270.007812 93.835938 267.535156 95.402344 C 265.0625 96.96875 262.195312 97.753906 258.9375 97.753906 Z M 258.9375 88.601562 C 260.09375 88.601562 261.082031 88.167969 261.90625 87.300781 C 262.730469 86.4375 263.371094 85.136719 263.824219 83.40625 C 264.277344 81.632812 264.503906 79.40625 264.503906 76.726562 C 264.503906 74.171875 264.277344 72.046875 263.824219 70.355469 C 263.371094 68.625 262.730469 67.347656 261.90625 66.523438 C 261.082031 65.65625 260.09375 65.222656 258.9375 65.222656 C 257.785156 65.222656 256.792969 65.65625 255.96875 66.523438 C 255.144531 67.347656 254.507812 68.625 254.054688 70.355469 C 253.597656 72.046875 253.371094 74.171875 253.371094 76.726562 C 253.371094 79.40625 253.597656 81.632812 254.054688 83.40625 C 254.507812 85.136719 255.144531 86.4375 255.96875 87.300781 C 256.792969 88.167969 257.785156 88.601562 258.9375 88.601562 Z M 281.179688 97.011719 L 281.179688 56.8125 L 291.9375 56.8125 L 298.679688 74.746094 L 301.957031 83.777344 L 302.207031 83.777344 C 301.957031 81.589844 301.648438 79.117188 301.277344 76.355469 C 300.90625 73.550781 300.71875 70.914062 300.71875 68.4375 L 300.71875 56.8125 L 310.863281 56.8125 L 310.863281 97.011719 L 300.101562 97.011719 L 293.363281 79.015625 L 290.082031 70.046875 L 289.835938 70.046875 C 290.082031 72.316406 290.390625 74.828125 290.765625 77.59375 C 291.136719 80.3125 291.320312 82.910156 291.320312 85.382812 L 291.320312 97.011719 Z M 327.785156 97.011719 L 327.785156 65.71875 L 316.21875 65.71875 L 316.21875 56.8125 L 349.988281 56.8125 L 349.988281 65.71875 L 338.421875 65.71875 L 338.421875 97.011719 Z M 356.828125 97.011719 L 356.828125 56.8125 L 384.164062 56.8125 L 384.164062 65.71875 L 367.464844 65.71875 L 367.464844 72.027344 L 381.6875 72.027344 L 381.6875 80.933594 L 367.464844 80.933594 L 367.464844 88.105469 L 384.78125 88.105469 L 384.78125 97.011719 Z M 392.425781 97.011719 L 392.425781 56.8125 L 403.1875 56.8125 L 409.929688 74.746094 L 413.203125 83.777344 L 413.453125 83.777344 C 413.203125 81.589844 412.894531 79.117188 412.523438 76.355469 C 412.15625 73.550781 411.96875 70.914062 411.96875 68.4375 L 411.96875 56.8125 L 422.113281 56.8125 L 422.113281 97.011719 L 411.351562 97.011719 L 404.609375 79.015625 L 401.332031 70.046875 L 401.085938 70.046875 C 401.332031 72.316406 401.640625 74.828125 402.011719 77.59375 C 402.382812 80.3125 402.566406 82.910156 402.566406 85.382812 L 402.566406 97.011719 Z M 429.753906 97.011719 L 429.753906 56.8125 L 441.320312 56.8125 C 445.152344 56.8125 448.535156 57.492188 451.460938 58.855469 C 454.429688 60.171875 456.742188 62.296875 458.390625 65.222656 C 460.078125 68.152344 460.925781 71.984375 460.925781 76.726562 C 460.925781 81.46875 460.101562 85.34375 458.449219 88.351562 C 456.800781 91.320312 454.535156 93.507812 451.648438 94.910156 C 448.800781 96.308594 445.566406 97.011719 441.9375 97.011719 Z M 440.390625 88.476562 L 440.765625 88.476562 C 442.496094 88.476562 444.058594 88.148438 445.464844 87.488281 C 446.867188 86.828125 447.980469 85.652344 448.800781 83.960938 C 449.628906 82.273438 450.039062 79.859375 450.039062 76.726562 C 450.039062 73.59375 449.628906 71.222656 448.800781 69.613281 C 447.980469 67.964844 446.867188 66.851562 445.464844 66.273438 C 444.058594 65.65625 442.496094 65.347656 440.765625 65.347656 L 440.390625 65.347656 Z M 481.433594 97.753906 C 479.042969 97.753906 477.101562 96.96875 475.621094 95.402344 C 474.132812 93.835938 473.394531 91.941406 473.394531 89.714844 C 473.394531 87.527344 474.132812 85.652344 475.621094 84.085938 C 477.101562 82.476562 479.042969 81.675781 481.433594 81.675781 C 483.863281 81.675781 485.804688 82.476562 487.246094 84.085938 C 488.730469 85.652344 489.472656 87.527344 489.472656 89.714844 C 489.472656 91.941406 488.730469 93.835938 487.246094 95.402344 C 485.804688 96.96875 483.863281 97.753906 481.433594 97.753906 Z M 503.179688 97.011719 L 503.179688 56.8125 L 514 56.8125 L 517.277344 69.921875 L 518.390625 75.117188 L 518.640625 75.117188 L 519.753906 69.921875 L 523.027344 56.8125 L 533.851562 56.8125 L 533.851562 97.011719 L 525.070312 97.011719 L 525.070312 87.550781 C 525.070312 86.355469 525.113281 84.847656 525.195312 83.035156 C 525.320312 81.21875 525.441406 79.324219 525.566406 77.34375 C 525.6875 75.324219 525.832031 73.40625 526 71.59375 C 526.164062 69.78125 526.308594 68.296875 526.429688 67.140625 L 526.246094 67.140625 L 522.84375 79.078125 L 520.742188 86.375 L 516.039062 86.375 L 513.816406 79.078125 L 510.722656 67.140625 L 510.476562 67.140625 C 510.597656 68.296875 510.742188 69.78125 510.910156 71.59375 C 511.070312 73.40625 511.21875 75.324219 511.34375 77.34375 C 511.464844 79.324219 511.570312 81.21875 511.648438 83.035156 C 511.773438 84.847656 511.835938 86.355469 511.835938 87.550781 L 511.835938 97.011719 Z M 555.722656 97.753906 C 553.082031 97.753906 550.816406 97.382812 548.917969 96.640625 C 547.019531 95.855469 545.453125 94.765625 544.21875 93.363281 C 543.023438 91.917969 542.117188 90.207031 541.496094 88.230469 C 540.917969 86.210938 540.632812 83.941406 540.632812 81.425781 L 540.632812 56.8125 L 551.269531 56.8125 L 551.269531 82.664062 C 551.269531 83.816406 551.394531 84.847656 551.640625 85.753906 C 551.929688 86.621094 552.402344 87.324219 553.0625 87.859375 C 553.722656 88.351562 554.609375 88.601562 555.722656 88.601562 C 556.875 88.601562 557.78125 88.351562 558.441406 87.859375 C 559.101562 87.324219 559.578125 86.621094 559.863281 85.753906 C 560.152344 84.847656 560.296875 83.816406 560.296875 82.664062 L 560.296875 56.8125 L 570.566406 56.8125 L 570.566406 81.425781 C 570.566406 83.941406 570.273438 86.210938 569.699219 88.230469 C 569.160156 90.207031 568.296875 91.917969 567.101562 93.363281 C 565.90625 94.765625 564.359375 95.855469 562.464844 96.640625 C 560.605469 97.382812 558.359375 97.753906 555.722656 97.753906 Z M 555.722656 97.753906 "
-					fill-opacity="1"
-					fill-rule="nonzero"
+					fillOpacity="1"
+					fillRule="nonzero"
 				/>
 			</svg>
 		</div>
