+++ src/components/LogoMinimal.tsx
@@ -23,29 +23,29 @@ export const LogoMinimal: React.FC<{
 					<clipPath id="4ac60d2535">
 						<path
 							d="M 185.570312 142.863281 L 703.707031 142.863281 L 703.707031 607.332031 L 185.570312 607.332031 Z M 185.570312 142.863281 "
-							clip-rule="nonzero"
+							clipRule="nonzero"
 						/>
 					</clipPath>
 					<clipPath id="1c73f312de">
 						<path
 							d="M 185.570312 146 L 698 146 L 698 602 L 185.570312 602 Z M 185.570312 146 "
-							clip-rule="nonzero"
+							clipRule="nonzero"
 						/>
 					</clipPath>
 					<clipPath id="068d4c4bb3">
 						<path
 							d="M 70.582031 70.582031 L 739.414062 70.582031 L 739.414062 739.417969 L 70.582031 739.417969 Z M 70.582031 70.582031 "
-							clip-rule="nonzero"
+							clipRule="nonzero"
 						/>
 					</clipPath>
 					<clipPath id="90f7f1637d">
 						<path
 							d="M 405 70.582031 C 220.308594 70.582031 70.582031 220.308594 70.582031 405 C 70.582031 589.691406 220.308594 739.414062 405 739.414062 C 589.691406 739.414062 739.414062 589.691406 739.414062 405 C 739.414062 220.308594 589.691406 70.582031 405 70.582031 Z M 405 70.582031 "
-							clip-rule="nonzero"
+							clipRule="nonzero"
 						/>
 					</clipPath>
 				</defs>
-				<g clip-path="url(#4ac60d2535)">
+				<g clipPath="url(#4ac60d2535)">
 					<path
 						strokeLinecap="butt"
 						transform="matrix(0.479692, 0, 0, 0.479692, 57.114104, -3.340133)"
@@ -55,19 +55,19 @@ export const LogoMinimal: React.FC<{
 						stroke="#1c6fbc"
 						strokeWidth="15.547"
 						strokeOpacity="1"
-						stroke-miterlimit="4"
+						strokeMiterlimit="4"
 					/>
 				</g>
-				<g clip-path="url(#1c73f312de)">
+				<g clipPath="url(#1c73f312de)">
 					<path
 						fill={color}
 						d="M 377.824219 146.238281 L 459.273438 146.238281 L 358.476562 383.75 L 358.457031 383.800781 L 266.863281 601.628906 L 185.4375 601.628906 Z M 506.769531 194.742188 L 697.027344 385.046875 L 353.695312 557.511719 L 397.683594 452.707031 L 575.480469 368.535156 L 475.742188 267.382812 Z M 506.769531 194.742188 "
-						fill-opacity="1"
-						fill-rule="evenodd"
+						fillOpacity="1"
+						fillRule="evenodd"
 					/>
 				</g>
-				<g clip-path="url(#068d4c4bb3)">
-					<g clip-path="url(#90f7f1637d)">
+				<g clipPath="url(#068d4c4bb3)">
+					<g clipPath="url(#90f7f1637d)">
 						<path
 							strokeLinecap="butt"
 							transform="matrix(2.064296, 0, 0, 2.064296, 70.583964, 70.58398)"
@@ -77,7 +77,7 @@ export const LogoMinimal: React.FC<{
 							stroke="#ffffff"
 							strokeWidth="2.906559"
 							strokeOpacity="1"
-							stroke-miterlimit="4"
+							strokeMiterlimit="4"
 						/>
 					</g>
 				</g>
