+++ src/index.ts
@@ -1,4 +1,4 @@
 import {registerRoot} from 'remotion';
-import {RemotionRoot} from './Root';
+import {RemotionRoot} from './components/Root';
 
 registerRoot(RemotionRoot);
